fixed gnu warnings

+++ src/ca/oldCAC.cpp
@@ -153,7 +153,9 @@ void oldCAC::exception ( int status, const char *pContext,
     else {
         this->clientCtx.signal ( status, pFileName, lineNo, 
             "op=%u, channel=%s, type=%s, count=lu, ctx=\"%s\"", 
-            op, ca_name ( &chan ), dbr_type_to_text ( type ), count, pContext );
+            op, ca_name ( &chan ), 
+            dbr_type_to_text ( static_cast <int> ( type ) ), 
+            count, pContext );
     }
 }
